Add cookies and compression

+++ index.js
@@ -12,6 +12,39 @@ import path from 'path'
 import url from 'url'
 import { randomUUID } from 'crypto'
 import { fork } from 'child_process'
+import { brotliCompress } from 'zlib'
+
+const cookieRecepie = () => ({ id: randomUUID(), value: randomUUID() })
+class CookieJar {
+  #cookies = new Map()
+  set(id, cookie) {
+    if (cookie.id) {
+      this.#cookies.set(id, cookie)
+      setTimeout(
+        () => this.#cookies.delete(id),
+        this.#cookies.get(id).maxAge * 1000
+      )
+    }
+  }
+  get(id) {
+    return this.#cookies.get(id)
+  }
+  destroy(id) {
+    this.#cookies.delete(id)
+  }
+  keys() {
+    return this.#cookies.keys()
+  }
+  values() {
+    return [...this.#cookies.values()]
+  }
+  isCookieVerified([id, value]) {
+    const current = this.get(id)
+    return current && value === current.value
+  }
+}
+
+const cookieJar = new CookieJar()
 
 const runScript = (scriptPath, args, callback) => {
   let invoked = false
@@ -66,38 +99,54 @@ const getReqData = (req) =>
   })
 
 const handleChanges = (data = [], buffer = '') => {
-  const characters = buffer.split('')
-  const result = []
+  const characters = buffer
   let pointer = 0
-  data.forEach((change) => {
+  return data.reduce((result, change) => {
     if (change[0] === 0) {
-      for (let i = pointer; i < pointer + change[1]; i++) {
-        result.push(characters[i])
-      }
+      for (let i = pointer; i < pointer + change[1]; ++i)
+        result += characters[i]
       pointer += change[1]
-    } else if (change[0] === -1) {
-      pointer += change[1]
-    } else if (change[0] === 1) {
-      result.push(...change[1])
-    }
-  })
-  return result.join('')
+    } else if (change[0] === -1) pointer += change[1]
+    else if (change[0] === 1) result += change[1]
+    return result
+  }, '')
 }
 
 const router = {
   GET: {},
   POST: {},
+  PUT: {},
   DELETE: {},
 }
 
-router['GET']['/dir'] = async (req, res, { query }) => {
-  const uuid = randomUUID()
-  const dir = directoryName + '/portals/' + uuid
+router['GET']['/dir'] = async (req, res) => {
+  const creds = cookieRecepie()
+  const dir = directoryName + '/portals/' + creds.id
+  const maxAge = 60 * 30
+  const cookie = {
+    id: creds.id,
+    value: creds.value,
+    maxAge,
+  }
   mkdir(dir)
+  res.setHeader(
+    'Set-Cookie',
+    `Value=${creds.id}.${creds.value}; Max-Age=${maxAge}; Path=/; HttpOnly; SameSite=Strict; Secure`
+  )
+  cookieJar.set(creds.id, cookie)
   res.writeHead(200, { 'Content-Type': 'application/text' })
-  res.end(uuid)
+  // res.headers(
+  //   'Set-Cookie',
+  //   `Max-Age=${maxAge}; Path=/; HttpOnly; SameSite=Strict`
+  // )
+  res.end(creds.id)
 }
-router['GET']['/ls'] = async (req, res, { query }) => {
+router['GET']['/ls'] = async (req, res, { query, cookie }) => {
+  if (!cookieJar.isCookieVerified(cookie)) {
+    res.writeHead(403, { 'Content-Type': 'text/html' })
+    res.end('403: Unauthorized!')
+    return
+  }
   const dir = directoryName + '/portals/' + query.dir + '/'
   await access(dir, constants.F_OK)
     .then(async () => {
@@ -106,11 +155,16 @@ router['GET']['/ls'] = async (req, res, { query }) => {
       res.end(JSON.stringify(list))
     })
     .catch((err) => {
-      res.writeHead(404, { 'Content-Type': 'text/html' })
-      res.end('404: Folder not found')
+      res.writeHead(404, { 'Content-Type': 'application/json' })
+      res.end('[]')
     })
 }
-router['POST']['/save'] = async (req, res, { query }) => {
+router['POST']['/save'] = async (req, res, { query, cookie }) => {
+  if (!cookieJar.isCookieVerified(cookie)) {
+    res.writeHead(403, { 'Content-Type': 'text/html' })
+    res.end('403: Unauthorized!')
+    return
+  }
   const data = JSON.parse(await getReqData(req))
   const dir = `${directoryName}/portals/${query.dir}/`
   const filepath = `${dir}${query.filename}`
@@ -123,6 +177,11 @@ router['POST']['/save'] = async (req, res, { query }) => {
         })
         .catch(async () => {
           const file = handleChanges(data, '')
+          const folders = query.filename.split('/')
+          folders.pop()
+          if (folders.length) {
+            await mkdir(`${dir}/${folders.join('/')}`, { recursive: true })
+          }
           await writeFile(filepath, file)
         })
     })
@@ -131,17 +190,24 @@ router['POST']['/save'] = async (req, res, { query }) => {
   res.writeHead(200, { 'Content-Type': 'application/text' })
   res.end()
 }
-router['POST']['/exec'] = async (req, res, { query }) => {
+router['POST']['/exec'] = async (req, res, { query, cookie }) => {
+  if (!cookieJar.isCookieVerified(cookie)) {
+    res.writeHead(403, { 'Content-Type': 'text/html' })
+    res.end('403: Unauthorized!')
+    return
+  }
   const dir = `${directoryName}/portals/${query.dir}/`
   const filepath = `${dir}${query.filename}`
-  runScript(filepath, [dir], (err) => {
-    if (err) return console.log(err)
-    console.log('finished running ' + filepath)
-  })
+  runScript(filepath, [dir], (err) => err && console.log(err))
   res.writeHead(200, { 'Content-Type': 'application/text' })
   res.end()
 }
-router['POST']['/disconnect'] = async (req, res, { query }) => {
+router['POST']['/disconnect'] = async (req, res, { query, cookie }) => {
+  if (!cookieJar.isCookieVerified(cookie)) {
+    res.writeHead(403, { 'Content-Type': 'text/html' })
+    res.end('403: Unauthorized!')
+    return
+  }
   const filepath = `${directoryName}/portals/${query.dir}`
   access(filepath, constants.F_OK)
     .then(() => rm(filepath, { recursive: true }, () => {}))
@@ -149,13 +215,31 @@ router['POST']['/disconnect'] = async (req, res, { query }) => {
   res.writeHead(200, { 'Content-Type': 'application/json' })
   res.end()
 }
-router['DELETE']['/del'] = async (req, res, { query }) => {
+router['DELETE']['/del'] = async (req, res, { query, cookie }) => {
+  if (!cookieJar.isCookieVerified(cookie)) {
+    res.writeHead(403, { 'Content-Type': 'text/html' })
+    res.end('403: Unauthorized!')
+    return
+  }
   const filepath = `${directoryName}/portals/${query.dir}/${query.filename}`
   access(filepath, constants.F_OK).then(() => unlink(filepath))
   res.writeHead(200, { 'Content-Type': 'application/json' })
   res.end()
 }
-router['GET']['*'] = async (req, res, { query, pathname }) => {
+router['DELETE']['/empty'] = async (req, res, { query, cookie }) => {
+  if (!cookieJar.isCookieVerified(cookie)) {
+    res.writeHead(403, { 'Content-Type': 'text/html' })
+    res.end('403: Unauthorized!')
+    return
+  }
+  const dir = `${directoryName}/portals/${query.dir}/`
+  access(dir, constants.F_OK).then(() =>
+    rm(dir, { recursive: true }, () => mkdir(dir))
+  )
+  res.writeHead(200, { 'Content-Type': 'application/json' })
+  res.end()
+}
+router['GET']['*'] = async (req, res, { pathname }) => {
   const extension = path.extname(req.url).slice(1)
   const type = extension ? types[extension] : types.html
   const supportedExtension = Boolean(type)
@@ -180,8 +264,11 @@ router['GET']['*'] = async (req, res, { query, pathname }) => {
     return
   }
   try {
+    res.setHeader('Content-Encoding', 'br')
     res.writeHead(200, { 'Content-Type': type })
-    res.end(await readFile(filePath))
+    brotliCompress(await readFile(filePath), {}, (error, buffer) =>
+      error ? console.log(error) : res.end(buffer)
+    )
   } catch (err) {
     res.writeHead(404, { 'Content-Type': 'text/html' })
     res.end('404: File not found')
@@ -199,20 +286,21 @@ const match = (method, pathname, req, res, params) => {
 const server = http.createServer(async (req, res) => {
   const URL = url.parse(req.url, true)
   const { query, pathname } = URL
-  const params = { query, pathname }
-  match(req.method, pathname, req, res, params) ??
-    match('GET', '*', req, res, params)
+  const cookie = req.headers.cookie?.split('Value=')[1].split('.')
+  match(req.method, pathname, req, res, {
+    query,
+    pathname,
+    cookie,
+  }) ?? match('GET', '*', req, res, { pathname })
 })
 
-server.listen(PORT, () => {
+server.listen(PORT, () =>
   access(root + '/portals', constants.F_OK)
-    .then(async () => {
+    .then(async () =>
       rm(root + '/portals', { recursive: true }, (err) =>
         mkdirSync(root + '/portals', (err) => err)
       )
-    })
-    .catch(() => {
-      mkdirSync(root + '/portals', (err) => err)
-    })
-  console.log(`server started on port: ${PORT}`)
-})
+    )
+    .catch(() => mkdirSync(root + '/portals', (err) => err))
+    .finally(() => console.log(`server started on port: ${PORT}`))
+)
